refactor: extracted components
