+++ apps/web/src/app/[locale]/(protected)/monster/types.ts
@@ -1,0 +1,4 @@
+export type Monster = {
+    name: string;
+    image: string;
+};
